adding more structure to files, created a main-content container and moved smaller navbar components to another file
