+++ client/src/App.css
@@ -13,12 +13,6 @@ a {
   color: unset;
   text-decoration:underline;
 }
-
-.gradient-bg {
-  /* global 92%+ browsers support */
-  background: radial-gradient(circle at 10% 10%, rgba(255, 255, 255, 1) 0%, rgba(179, 226, 248, 1) 100%, rgba(179, 226, 248, 1) 100%);
-}
-
 .gradient-text {
   background: var(--color-bg);
 }
@@ -29,32 +23,6 @@ a {
 
 .section__margin {
   margin: 4rem 6rem;
-}
-
-.scale-up-center {
-	-webkit-animation: scale-up-center 1000ms cubic-bezier(0.390, 0.575, 0.565, 1.000) both;
-	        animation: scale-up-center 1000ms cubic-bezier(0.390, 0.575, 0.565, 1.000) both;
-}
-
-@-webkit-keyframes scale-up-center {
-0% {
-  -webkit-transform: scale(0.5);
-          transform: scale(0.5);
-}
-100% {
-  -webkit-transform: scale(1);
-          transform: scale(1);
-}
-}
-@keyframes scale-up-center {
-  0% {
-    -webkit-transform: scale(0.5);
-            transform: scale(0.5);
-  }
-  100% {
-    -webkit-transform: scale(1);
-            transform: scale(1);
-  }
 }
 
 /* medium screen devices */
