+++ client/src/components/navbar/smaller-components/Sections.jsx
@@ -1,6 +1,6 @@
-import './navbar.css'
+import '../navbar.css'
 
-export function Sections() {
+export default function Sections() {
     return (
         <>
           <div className="navbar-links">
