+++ client/src/components/page-content/PageContent.jsx
@@ -1,0 +1,13 @@
+import "./page-content.css";
+
+function MainContent() {
+    return (
+        <>
+            <div className="aboutus">
+                asdasd
+            </div>
+        </>
+    )
+}
+
+export default MainContent
