+++ client/src/containers/footer/Footer.jsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import './footer.css'
 
 function Footer() {
