+++ client/src/containers/main-content/Main.jsx
@@ -1,0 +1,8 @@
+import MainContent from "../../components/page-content/PageContent";
+import "./main.css";
+
+function Main() {
+    return (
+        <MainContent />
+    )
+}
